Add ability to travel to location from dashbaord
